feat: hide columns cell override (#185)
We have the requirement to hide a column in a specific scenario. I've added the possibility to hide columns while in cells view for the iModelGrid.

I've added an enum that represents each columns in cell view.

I've also updated the storybook so that we can play with that option.


Co-authored-by: Arun George <11051042+aruniverse@users.noreply.github.com>