feat: add option to disable favorite iModels fetching (#189)
Given that we can hide the favorite icon and the favorite column, we introduce a new option so that we can disable fetching favorite iModels when the user does not care about the feature.

This prevents doing unnecessary calls to the favorites endpoint. This can also prevent unwanted errors when the user does not have access to the favorites endpoint (i.e.: custom shared contract)